requirements.txt
pandas==1.4.1
python-dotenv==0.19.2
requests==2.27.1
yagmail==0.15.277
